refactor(460): 메인 상단 biasList prefetch 구현
